requirements.txt
Django==1.6.5
coverage==3.7.1
django-coverage==1.2.4
py-trello==0.2.3
South==1.0
